+++ src/app.ts
@@ -1,360 +1,12 @@
-/// <reference path="drag-drop-interfaces.ts" />
-/// <reference path="project-model.ts" />
+/// <reference path="models/drag-drop.ts" />
+/// <reference path="models/project.ts" />
+/// <reference path="state/project-state.ts" />
+/// <reference path="util/validation.ts" />
+/// <reference path="decorators/autobind.ts" />
+/// <reference path="components/project-input.ts" />
+/// <reference path="components/project-list.ts" />
 
 namespace App {
-  // Project State Management
-  type listener<T> = (items: T[]) => void;
-
-  class State<T> {
-    protected listeners: listener<T>[] = [];
-    addListener(linstenerFn: listener<T>) {
-      this.listeners.push(linstenerFn);
-    }
-  }
-
-  class ProjectState extends State<Project> {
-    private projects: Project[] = [];
-    private static instance: ProjectState;
-
-    private constructor() {
-      super();
-    }
-
-    static getInstance() {
-      if (this.instance) {
-        return this.instance;
-      }
-      this.instance = new ProjectState();
-      return this.instance;
-    }
-
-    addProject(title: string, description: string, manday: number) {
-      const newProject = new Project(
-        Math.random().toString(),
-        title,
-        description,
-        manday,
-        ProjectStatus.Active
-      );
-      this.projects.push(newProject);
-      this.updateListeners();
-    }
-
-    moveProject(projectId: string, newStatus: ProjectStatus) {
-      const project = this.projects.find((prj) => prj.id === projectId);
-      if (project && project.status !== newStatus) {
-        project.status = newStatus;
-        this.updateListeners();
-      }
-    }
-
-    private updateListeners() {
-      for (const linstenerFn of this.listeners) {
-        linstenerFn(this.projects.slice());
-      }
-    }
-  }
-
-  const projectState = ProjectState.getInstance();
-
-  // validation
-  interface Validatable {
-    value?: string | number;
-    required?: boolean;
-    minLength?: number;
-    maxLength?: number;
-    min?: number;
-    max?: number;
-  }
-
-  function validate(validatableInput: Validatable) {
-    let isValid = true;
-    if (validatableInput.required) {
-      isValid =
-        isValid && validatableInput.value?.toString().trim().length !== 0;
-    }
-    if (
-      validatableInput.minLength !== undefined &&
-      typeof validatableInput.value === "string"
-    ) {
-      isValid =
-        isValid && validatableInput.value.length >= validatableInput.minLength;
-    }
-    if (
-      validatableInput.maxLength !== undefined &&
-      typeof validatableInput.value === "string"
-    ) {
-      isValid =
-        isValid && validatableInput.value.length <= validatableInput.maxLength;
-    }
-    if (
-      validatableInput.min !== undefined &&
-      typeof validatableInput.value === "number"
-    ) {
-      isValid = isValid && validatableInput.value >= validatableInput.min;
-    }
-    if (
-      validatableInput.max !== undefined &&
-      typeof validatableInput.value === "number"
-    ) {
-      isValid = isValid && validatableInput.value <= validatableInput.max;
-    }
-    return isValid;
-  }
-
-  // autobind decorator
-  function autobind(_: any, _2: string, descriptor: PropertyDescriptor) {
-    const originalMethod = descriptor.value;
-    const adjDescriptor: PropertyDescriptor = {
-      configurable: true,
-      get() {
-        const boundFn = originalMethod.bind(this);
-        return boundFn;
-      },
-    };
-    return adjDescriptor;
-  }
-
-  // Component Class
-  abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-    templateElement: HTMLTemplateElement; // template要素取得用
-    hostElement: T; // template要素出力用
-    element: U; // template内要素取得用
-
-    constructor(
-      templateId: string,
-      hostElementId: string,
-      insertAtStart: boolean,
-      newElementId?: string
-    ) {
-      // tamplateタグの内容を取得
-      this.templateElement = document.getElementById(
-        templateId
-      )! as HTMLTemplateElement;
-      this.hostElement = document.getElementById(hostElementId)! as T;
-
-      const importedNode = document.importNode(
-        this.templateElement.content,
-        true
-      );
-      this.element = importedNode.firstElementChild as U;
-      if (newElementId) {
-        this.element.id = newElementId;
-      }
-      this.attach(insertAtStart);
-    }
-
-    abstract configure(): void;
-    abstract renderContent(): void;
-
-    private attach(insertAtBeginning: boolean) {
-      this.hostElement.insertAdjacentElement(
-        insertAtBeginning ? "afterbegin" : "beforeend",
-        this.element
-      );
-    }
-  }
-
-  // ProjectItem Class
-  class ProjectItem
-    extends Component<HTMLUListElement, HTMLLIElement>
-    implements Draggable
-  {
-    private project: Project;
-
-    get manday() {
-      if (this.project.manday < 20) {
-        return this.project.manday.toString() + "人日";
-      } else {
-        return (this.project.manday / 20).toString() + "人月";
-      }
-    }
-
-    constructor(hostId: string, project: Project) {
-      super("single-project", hostId, false, project.id);
-      this.project = project;
-
-      this.configure();
-      this.renderContent();
-    }
-    @autobind
-    dragStartHandler(event: DragEvent): void {
-      event.dataTransfer!.setData("text/plain", this.project.id);
-      event.dataTransfer!.effectAllowed = "move";
-    }
-    dragEndHandler(_: DragEvent): void {
-      console.log("Drag終了");
-    }
-
-    configure(): void {
-      this.element.addEventListener("dragstart", this.dragStartHandler);
-      this.element.addEventListener("dragend", this.dragEndHandler);
-    }
-
-    renderContent(): void {
-      this.element.querySelector("h2")!.textContent = this.project.title;
-      this.element.querySelector("h3")!.textContent = this.manday;
-      this.element.querySelector("p")!.textContent = this.project.description;
-    }
-  }
-
-  // ProjectList Class
-  class ProjectList
-    extends Component<HTMLDivElement, HTMLElement>
-    implements DragTarget
-  {
-    assigneProjects: Project[];
-
-    constructor(private type: "active" | "finished") {
-      super("project-list", "app", false, `${type}-projects`);
-      this.assigneProjects = [];
-
-      this.configure();
-      this.renderContent();
-    }
-
-    @autobind
-    dragOverHandler(event: DragEvent): void {
-      if (event.dataTransfer && event.dataTransfer.types[0] === "text/plain") {
-        event.preventDefault();
-        const listEl = this.element.querySelector("ul")!;
-        listEl.classList.add("droppable");
-      }
-    }
-
-    @autobind
-    dropHandler(event: DragEvent): void {
-      const prjId = event.dataTransfer!.getData("text/plain");
-      projectState.moveProject(
-        prjId,
-        this.type === "active" ? ProjectStatus.Active : ProjectStatus.Finished
-      );
-    }
-
-    @autobind
-    dragLeaveHandler(_: DragEvent): void {
-      const listEl = this.element.querySelector("ul")!;
-      listEl.classList.remove("droppable");
-    }
-
-    configure() {
-      this.element.addEventListener("dragover", this.dragOverHandler);
-      this.element.addEventListener("drop", this.dropHandler);
-      this.element.addEventListener("dragleave", this.dragLeaveHandler);
-
-      projectState.addListener((projects: Project[]) => {
-        const relevantProjects = projects.filter((prj) => {
-          if (this.type === "active") {
-            return prj.status === ProjectStatus.Active;
-          }
-          return prj.status === ProjectStatus.Finished;
-        });
-        this.assigneProjects = relevantProjects;
-        this.renderProjects();
-      });
-      this.renderContent();
-    }
-
-    renderContent() {
-      const listId = `${this.type}-projects-list`;
-      this.element.querySelector("ul")!.id = listId;
-      this.element.querySelector("h2")!.textContent =
-        this.type === "active" ? "実行中プロジェクト" : "完了プロジェクト";
-    }
-
-    private renderProjects() {
-      const listEl = document.getElementById(
-        `${this.type}-projects-list`
-      )! as HTMLUListElement;
-      listEl.innerHTML = "";
-      for (const prjItem of this.assigneProjects) {
-        new ProjectItem(listEl.id, prjItem);
-      }
-    }
-  }
-
-  // ProjectInput Class
-  class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
-    titleInputElement: HTMLInputElement; // form_title取得用
-    descriptionInputElement: HTMLInputElement; // form_description取得用
-    mandayInputElement: HTMLInputElement; // form_manday取得用
-
-    constructor() {
-      super("project-input", "app", true, "user-input");
-
-      // form内要素の入力値を取得
-      this.titleInputElement = this.element.querySelector(
-        "#title"
-      ) as HTMLInputElement;
-      this.descriptionInputElement = this.element.querySelector(
-        "#description"
-      ) as HTMLInputElement;
-      this.mandayInputElement = this.element.querySelector(
-        "#manday"
-      ) as HTMLInputElement;
-
-      this.configure();
-    }
-
-    // submitが押されたことを受信する
-    configure() {
-      this.element.addEventListener("submit", this.submitHandler);
-    }
-
-    renderContent() {}
-
-    private gatherUserInput(): [string, string, number] | void {
-      const enteredTitle = this.titleInputElement.value;
-      const enteredDescription = this.descriptionInputElement.value;
-      const enteredManday = this.mandayInputElement.value;
-
-      const titleValidatable: Validatable = {
-        value: enteredTitle,
-        required: true,
-      };
-      const descriptionValidatable: Validatable = {
-        value: enteredDescription,
-        required: true,
-        minLength: 5,
-      };
-      const mandayValidatable: Validatable = {
-        value: +enteredManday,
-        required: true,
-        min: 1,
-        max: 1000,
-      };
-      if (
-        !validate(titleValidatable) ||
-        !validate(descriptionValidatable) ||
-        !validate(mandayValidatable)
-      ) {
-        alert("入力値が正しくありません。再度お試しください。");
-        return;
-      } else {
-        return [enteredTitle, enteredDescription, +enteredManday];
-      }
-    }
-
-    private clearInputs() {
-      this.titleInputElement.value = "";
-      this.descriptionInputElement.value = "";
-      this.mandayInputElement.value = "";
-    }
-
-    // 指定されたタグ内の値を出力する
-    @autobind
-    private submitHandler(event: Event) {
-      event.preventDefault();
-      console.log(this.titleInputElement.value);
-      const userInput = this.gatherUserInput();
-      if (Array.isArray(userInput)) {
-        const [title, desc, manday] = userInput;
-        projectState.addProject(title, desc, manday);
-        this.clearInputs();
-      }
-    }
-  }
-
   new ProjectInput();
   new ProjectList("active");
   new ProjectList("finished");
