+++ src/components/base-component.ts
@@ -1,0 +1,44 @@
+namespace App {
+  // Component Class
+  export abstract class Component<
+    T extends HTMLElement,
+    U extends HTMLElement
+  > {
+    templateElement: HTMLTemplateElement; // template要素取得用
+    hostElement: T; // template要素出力用
+    element: U; // template内要素取得用
+
+    constructor(
+      templateId: string,
+      hostElementId: string,
+      insertAtStart: boolean,
+      newElementId?: string
+    ) {
+      // tamplateタグの内容を取得
+      this.templateElement = document.getElementById(
+        templateId
+      )! as HTMLTemplateElement;
+      this.hostElement = document.getElementById(hostElementId)! as T;
+
+      const importedNode = document.importNode(
+        this.templateElement.content,
+        true
+      );
+      this.element = importedNode.firstElementChild as U;
+      if (newElementId) {
+        this.element.id = newElementId;
+      }
+      this.attach(insertAtStart);
+    }
+
+    abstract configure(): void;
+    abstract renderContent(): void;
+
+    private attach(insertAtBeginning: boolean) {
+      this.hostElement.insertAdjacentElement(
+        insertAtBeginning ? "afterbegin" : "beforeend",
+        this.element
+      );
+    }
+  }
+}
