+++ src/components/project-input.ts
@@ -1,0 +1,85 @@
+/// <reference path="base-component.ts" />
+
+namespace App {
+  // ProjectInput Class
+  export class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
+    titleInputElement: HTMLInputElement; // form_title取得用
+    descriptionInputElement: HTMLInputElement; // form_description取得用
+    mandayInputElement: HTMLInputElement; // form_manday取得用
+
+    constructor() {
+      super("project-input", "app", true, "user-input");
+
+      // form内要素の入力値を取得
+      this.titleInputElement = this.element.querySelector(
+        "#title"
+      ) as HTMLInputElement;
+      this.descriptionInputElement = this.element.querySelector(
+        "#description"
+      ) as HTMLInputElement;
+      this.mandayInputElement = this.element.querySelector(
+        "#manday"
+      ) as HTMLInputElement;
+
+      this.configure();
+    }
+
+    // submitが押されたことを受信する
+    configure() {
+      this.element.addEventListener("submit", this.submitHandler);
+    }
+
+    renderContent() {}
+
+    private gatherUserInput(): [string, string, number] | void {
+      const enteredTitle = this.titleInputElement.value;
+      const enteredDescription = this.descriptionInputElement.value;
+      const enteredManday = this.mandayInputElement.value;
+
+      const titleValidatable: Validatable = {
+        value: enteredTitle,
+        required: true,
+      };
+      const descriptionValidatable: Validatable = {
+        value: enteredDescription,
+        required: true,
+        minLength: 5,
+      };
+      const mandayValidatable: Validatable = {
+        value: +enteredManday,
+        required: true,
+        min: 1,
+        max: 1000,
+      };
+      if (
+        !validate(titleValidatable) ||
+        !validate(descriptionValidatable) ||
+        !validate(mandayValidatable)
+      ) {
+        alert("入力値が正しくありません。再度お試しください。");
+        return;
+      } else {
+        return [enteredTitle, enteredDescription, +enteredManday];
+      }
+    }
+
+    private clearInputs() {
+      this.titleInputElement.value = "";
+      this.descriptionInputElement.value = "";
+      this.mandayInputElement.value = "";
+    }
+
+    // 指定されたタグ内の値を出力する
+    @autobind
+    private submitHandler(event: Event) {
+      event.preventDefault();
+      console.log(this.titleInputElement.value);
+      const userInput = this.gatherUserInput();
+      if (Array.isArray(userInput)) {
+        const [title, desc, manday] = userInput;
+        projectState.addProject(title, desc, manday);
+        this.clearInputs();
+      }
+    }
+  }
+}
